requirements.txt
aws-cdk-lib>=2.180.0
constructs>=10.0.0
boto3>=1.36.0
requests>=2.32.0
feedparser>=6.0.0
aws-lambda-powertools[all]>=3.5.0
pytest>=8.0.0
moto>=5.0.0
